+++ Instagram/Model/APIData.swift
@@ -6,16 +6,46 @@
 //
 
 import Foundation
+import SwiftUI
 
-struct APIData: Decodable {
-    let results: [Characters]
+
+
+struct APIDataView: View {
+    @StateObject var viewModel = ViewModel()
+    
+    var body: some View {
+        NavigationView {
+            List {
+                ForEach(viewModel.characters) {character in
+                    HStack {
+                        AsyncImage(url: URL(string:character.image), content: { image in
+                            image
+                                .resizable()
+                                .frame(width: 70, height: 70)
+                        }, placeholder: {
+                            ProgressView()
+                        })
+                        
+                        Text(character.name)
+                    }
+                    
+                }
+                
+            }
+            .navigationTitle("Personajes")
+            .onAppear {
+                viewModel.loadData()
+            }
+        }
+    }
 }
 
-struct Characters: Decodable, Identifiable {
-    let id: Int
-    let name: String
-    let image: String
-}
 
+
+struct APIDataView_Previews: PreviewProvider {
+    static var previews: some View {
+        APIDataView()
+    }
+}
