+++ Instagram/View/ProfileView/ProfileView.swift
@@ -9,22 +9,22 @@ import SwiftUI
 
 struct ProfileView: View {
     
-    @State private var character: Characters = Characters(id: 1, name: "Jerry Smith", image: "https://rickandmortyapi.com/api/character/avatar/5.jpeg")
+    @StateObject var viewModel = ViewModel()
     
     var body: some View {
         VStack() {
-            ProfileTopHeader(characters: $character)
+            ProfileTopHeader(characters: viewModel.character)
             ScrollView {
                 VStack(alignment: .leading) {
-                    ProfileHeaderView(characters: $character)
+                    ProfileHeaderView(characters: viewModel.character)
                     
-                    Text(character.name).font(.title2).padding(4)
+                    Text(viewModel.character.name).font(.title2).padding(4)
                     
                     ProfileControlButtonView()
                     
                     HStack {
-                        HomeStoriesApi(characters: $character)
-                        HomeStoriesApi(characters: $character)
+                        HomeStoriesApi(characters: viewModel.character)
+                        HomeStoriesApi(characters: viewModel.character)
                     }
                     
                     ProfileMediaSelectionView()
@@ -35,29 +35,13 @@ struct ProfileView: View {
             }
         }
         .onAppear {
-            loadData()
-            
+            viewModel.singleLoadData()
         }
-    }
-    
-    private func loadData() {
-        guard let url = URL(string: "https://rickandmortyapi.com/api/character") else {
-            return
-        }
-        URLSession.shared.dataTask(with: url) { data, response, error in
-            guard let data = data else { return }
-            if let decodedData = try? JSONDecoder().decode(APIData.self, from: data) {
-                DispatchQueue.main.async {
-                    guard let character = decodedData.results.first else { return }
-                    self.character = character
-                }
-            }
-        }.resume()
     }
 }
 
 struct ProfileTopHeader: View {
-    @Binding var characters : Characters
+    @State var characters : Characters
     
     var body: some View {
         HStack {
